Major Refactor Changed all amounts/balances database field types from int to double Add Charge Date(chargeAt) field to Subscription to allows creation of recurring subscriptions and to also power the current impl of Charge

+++ lib/Insights/category.dart
@@ -22,7 +22,7 @@ class CategoryItem extends ConsumerWidget {
 
     final int spendsPerCategory =
         ref.watch(getSpendsByCategory(category.name)).length;
-    final int totalSpendAmountPerCategory =
+    final double totalSpendAmountPerCategory =
         ref.watch(getTotalSpendAmountPerCategory(category.name));
 
     final categoriesCount = ref.watch(categoriesNotifier).length;
